Developed the module
Fixed authentication completely, except verification of the API permissions. The authentication process is now fully managed in the back-end, and only Connect-CMPC is required to be run to authenticate the first time.
Fixed the import functions in the psm1 file.
Made some corrections in the psd1 file.
Worked on some of the remaining functions.
Got rid of one of the files in the Old folder by converting the script to new module format.