Open early beta test release
The open early beta test version is released.
Please remember to report bugs to nordby.mikael@gmail.com so they will be fixed.

Some not very necessary scripts are not done yet, as well as the Edit-CMPCAdminRelationship function, but they will be done before the release of the module.
Small bugs, unstandarized naming and other things are normal and expected.
In the Development folder, you can find bugs and improvements that are currently being fixed.